+++ api/migrations/0036_solid_goliath.sql
@@ -1,0 +1,2 @@
+ALTER TABLE "profile" ADD COLUMN "id" integer;--> statement-breakpoint
+ALTER TABLE "profile" ADD COLUMN "name" text DEFAULT 'Unknown';